Changing socket to file
