+++ kubernetes-entrypoint.go
@@ -8,9 +8,9 @@ import (
 	_ "github.com/stackanetes/kubernetes-entrypoint/dependencies/config"
 	_ "github.com/stackanetes/kubernetes-entrypoint/dependencies/container"
 	_ "github.com/stackanetes/kubernetes-entrypoint/dependencies/daemonset"
+	_ "github.com/stackanetes/kubernetes-entrypoint/dependencies/file"
 	_ "github.com/stackanetes/kubernetes-entrypoint/dependencies/job"
 	_ "github.com/stackanetes/kubernetes-entrypoint/dependencies/service"
-	_ "github.com/stackanetes/kubernetes-entrypoint/dependencies/socket"
 	"github.com/stackanetes/kubernetes-entrypoint/logger"
 	command "github.com/stackanetes/kubernetes-entrypoint/util/command"
 	"github.com/stackanetes/kubernetes-entrypoint/util/env"
